% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-res.R
\name{biasplot}
\alias{biasplot}
\title{energyplot}
\usage{
biasplot(
  stanres,
  dirname,
  filename,
  mat1,
  prof = prof,
  meansd = meansd,
  typesim = typesim,
  by = 5,
  pdf = F,
  ht = 10,
  wd = 10
)
}
\arguments{
\item{stanres}{Output from runstan}

\item{dirname}{Folder name where to save plots}

\item{filename}{Base filename}

\item{by}{Skips for parameter posterior draws (boxplot)}

\item{pdf}{Whether to save pdf, defaults to F}

\item{ht}{Height of pdf}

\item{wd}{Width of pdf}
}
\description{
\code{biasplot} Create energy plots for stan output
}
