% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa-sim-fn.R
\name{checkc2}
\alias{checkc2}
\title{checkc2}
\usage{
checkc2(dat)
}
\arguments{
\item{dat}{Data frame of NA (unconstrained), 0s, and 1s, consisting of a column of pollutants called poll and columns corresponding to constraints for each source}
}
\description{
\code{checkc2} Check rank condition C2 for Bayesian SA
}
