% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-res.R
\name{energyplot}
\alias{energyplot}
\title{energyplot}
\usage{
energyplot(stanres, dirname, filename, pdf = F, ht = 400, wd = 20)
}
\arguments{
\item{stanres}{Output from runstan}

\item{dirname}{Folder name where to save plots}

\item{filename}{Base filename}

\item{pdf}{Whether to save pdf, defaults to F}

\item{ht}{Height of pdf}

\item{wd}{Width of pdf}
}
\description{
\code{energyplot} Create energy plots for stan output
}
