% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstan.R
\name{findamb}
\alias{findamb}
\title{findamb}
\usage{
findamb(dat, amb)
}
\arguments{
\item{dat}{Simulated local data}

\item{amb}{Ambient data and stan results + stan data}

\item{profdat}{Profile data}

\item{meansd}{Matrix with columns corresponding to source, type, mean, sd}

\item{typesim}{Which simulation to run}

\item{N}{Number of days to simulate.  If not noninformative stan, vector of 2 for days in ambient, days in local}

\item{rmout}{Whether to remove largest}

\item{sderr}{Standard deviation of error.  If not noninformative stan, can be vector of 2 for ambient local}

\item{seed}{Seed for stan.  Must be a vector of length 2}
}
\description{
\code{findamb} Find ambient priors
}
