% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa-sim-fn.R
\name{genln}
\alias{genln}
\title{genln}
\usage{
genln(dat, N)
}
\arguments{
\item{dat}{Matrix with columns corresponding to mean, sd}

\item{N}{Number of days to simulate}
}
\description{
\code{genln} Generate lognormal data from dataset
}
