% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa-sim-fn.R
\name{genn}
\alias{genn}
\title{genn}
\usage{
genn(dat, N, log = T)
}
\arguments{
\item{dat}{Matrix with columns corresponding to mean, sd}

\item{N}{Number of days to simulate}

\item{log}{Whether lognormal}
}
\description{
\code{genn} Generate normal data from dataset
}
