% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-res.R
\name{getdat}
\alias{getdat}
\title{getdat}
\usage{
getdat(stanres, typeplot, L)
}
\arguments{
\item{typeplot}{Ambient or local}

\item{L}{number of sources}

\item{standat}{Stan data}
}
\description{
\code{getdat} Get ambient vs. local data
}
