% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstan.R
\name{getfmat}
\alias{getfmat}
\title{getfmat}
\usage{
getfmat(ffree, pos, onemat, zeromat)
}
\arguments{
\item{ffree}{Fmatrix of free elements from Stan output}

\item{pos}{position data}

\item{onemat}{Onemat constraints of 1}

\item{zeromat}{Zeromat constraints of 0s and 1s}
}
\description{
\code{getfmat} Get F matrix from Stan output of free elements
}
