% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-res.R
\name{makelabels}
\alias{makelabels}
\title{makelabels}
\usage{
makelabels(cons, sources, mat1)
}
\arguments{
\item{cons}{Names of pollutants in order}

\item{sources}{Names of sources in order}

\item{mat1}{Names of free elements in F in order}
}
\description{
\code{makelabels} Create names for variables
}
