% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-res.R
\name{mat1fun}
\alias{mat1fun}
\title{mat1fun}
\usage{
mat1fun(standat, sources)
}
\arguments{
\item{standat}{Stan data}

\item{sources}{Names of sources}
}
\description{
\code{mat1fun} Get constraints with labels
}
