% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meansd.R
\docType{data}
\name{meansd}
\alias{meansd}
\title{Mean and standard deviation of sources}
\format{
A dataframe with 19 rows and 4 variables
}
\description{
A dataset of PM2.5 source mean/SD in referent constituent
units (e.g., units of Silicon)
The variables are:
}
\details{
\itemize{
\item source. Source name
\item type. Simulation type (ambient, local1 - local4)
\item mean. Normal mean for each source
\item sd. Normal standard deviation for each source
...
}
}
\keyword{datasets}
