% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-res.R
\name{mytraceplot}
\alias{mytraceplot}
\title{mytraceplot}
\usage{
mytraceplot(stanres, dirname, filename, by = 5)
}
\arguments{
\item{stanres}{Output from runstan}

\item{dirname}{Folder name where to save plots}

\item{filename}{Base filename}

\item{by}{Skips for traceplot}
}
\description{
\code{mytraceplot} Create traceplots for stan output
}
