% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-res.R
\name{pairsplot}
\alias{pairsplot}
\title{pairsplot}
\usage{
pairsplot(
  stanres,
  dirname,
  filename,
  mat1,
  sources,
  cons,
  typeplot,
  cond1 = "energy"
)
}
\arguments{
\item{stanres}{Output from runstan}

\item{dirname}{Folder name where to save plots}

\item{filename}{Base filename}

\item{mat1}{Names of free elements of F}

\item{sources}{Names of sources}
}
\description{
\code{pairsplot} Create pairs plots for stan output
}
