% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa-sim-fn.R
\name{reorgout}
\alias{reorgout}
\title{reorgout}
\usage{
reorgout(outa, outl)
}
\arguments{
\item{outa}{Ambient data output from simdat1}

\item{outl}{Local data output from simdat1}
}
\description{
\code{reorgout} Reorganize local/ambient output
}
