% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-res.R
\name{rescoverage}
\alias{rescoverage}
\title{rescoverage}
\usage{
rescoverage(typesim, stanres, by = 5, prof = prof, meansd = meansd)
}
\arguments{
\item{typesim}{Which simulation to run.  Currently, local1 - local4, ambient.}

\item{stanres}{Output from runstan}

\item{by}{Skips for traceplot, biasplot}

\item{prof}{Profile data, defaults to prof.}

\item{meansd}{Matrix with columns corresponding to source, type, mean, sd, defaults to meansd.}
}
\description{
\code{rescoverage} Calculate coverage of results
}
