% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-res.R
\name{rhatplot}
\alias{rhatplot}
\title{rhatplot}
\usage{
rhatplot(stanres, dirname, filename, pdf = F, ht = 100)
}
\arguments{
\item{stanres}{Output from runstan}

\item{dirname}{Folder name where to save plots}

\item{filename}{Base filename}

\item{pdf}{Whether to save pdf, defaults to F}

\item{ht}{Height of pdf}
}
\description{
\code{rhatplot} Create rhat plots for stan output
}
