% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstan-data.R
\name{rundata}
\alias{rundata}
\title{rundata}
\usage{
rundata(data, stancode, seed, iter = 1000, chains = 1, keepall = F, ...)
}
\arguments{
\item{stancode}{Type of stan model to run}

\item{seed}{Seed for stan.  If inform, seed must be length 2}

\item{iter}{Number of iterations for MCMC}

\item{chains}{Number of chains}

\item{keepall}{Whether to keep all results or only summary}

\item{dat}{Data for stan model}
}
\description{
\code{rundata} Run stan for RWD
}
