% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstan.R
\name{runstan}
\alias{runstan}
\title{runstan}
\usage{
runstan(
  N,
  typesim,
  stancode = NULL,
  stantype = NULL,
  prof = prof,
  meansd = meansd,
  keep = "all",
  rmout = F,
  sderr = NULL,
  seeds = NULL,
  iter = 1000,
  chains = 1,
  findamb = NULL,
  notes = NULL,
  fp = NULL,
  names = F,
  ...
)
}
\arguments{
\item{N}{Number of days to simulate.  If not noninformative stan, vector of 2 for days in ambient, days in local}

\item{typesim}{Which simulation to run.  Currently, local1 - local4, ambient.}

\item{stancode}{Type of stan model to run}

\item{stantype}{Type of stan model (can differ from code: for ifelse statements)}

\item{prof}{Profile data, defaults to prof.}

\item{meansd}{Matrix with columns corresponding to source, type, mean, sd, defaults to meansd.}

\item{keep}{What to keep}

\item{rmout}{Whether to remove largest (outliers)}

\item{sderr}{Standard deviation of error.  If not noninformative stan, can be vector of 2 for ambient local}

\item{iter}{Number of iterations to run}

\item{chains}{Number of chains}

\item{findamb}{Results from find ambient for informative}

\item{seed}{Seeds for stan data simulation and running (length 2).  If inform, seed must be length 3}
}
\description{
\code{runstan} Simulate all data
}
