% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa-sim-fn.R
\name{simdat}
\alias{simdat}
\title{simdat}
\usage{
simdat(
  stanmodel,
  typesim,
  N,
  prof = prof,
  meansd = meansd,
  sderr = NULL,
  rmout = F
)
}
\arguments{
\item{stanmodel}{Type of stan model to run.  Currently only noninformative tested.}

\item{typesim}{Which simulation to run.  Currently, local1 - local4, ambient.}

\item{N}{Number of days to simulate.  If not noninformative stan, vector of 2 for days in ambient, days in local.}

\item{prof}{Profile data, defaults to prof}

\item{meansd}{Matrix with columns corresponding to source, type, mean, sd, defaults to meansd.}

\item{sderr}{Standard deviation of error. If not noninformative stan, optional vector of 2 for error in ambient, error in local.}

\item{rmout}{Whether to remove largest (outliers)}
}
\description{
\code{simdat} Simulate all data
}
