% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa-sim-fn.R
\name{simdat1}
\alias{simdat1}
\title{simdat1}
\usage{
simdat1(
  typesim = "ambient",
  N = 100,
  prof0 = prof,
  meansd0 = meansd,
  sderr = NULL,
  rmout = F,
  log1 = T
)
}
\arguments{
\item{typesim}{Which simulation to run.  Currently, local1 - local4, ambient.}

\item{N}{Number of days to simulate.}

\item{sderr}{Standard deviation of error.}

\item{rmout}{Whether to remove largest (outliers)}

\item{log1}{Whether to use lognormal for G}

\item{prof}{Profile data, defaults to prof.}

\item{meansd}{Matrix with columns corresponding to source, type, mean, sd, defaults to meansd.}
}
\description{
\code{simdat1} Inner functions to simulate each data (e.g., ambient or local)
}
