% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstan-data.R
\name{tblchain}
\alias{tblchain}
\title{tblchain}
\usage{
tblchain(x, by1 = 1, start = 5000)
}
\arguments{
\item{x}{data from one chain of stan, i.e. stanfitATsim$samples list #1}

\item{by1}{Skips for posterior draws}

\item{start}{Where to start by}
}
\description{
\code{tblchain} Format stan output for 1 chain as tibble
}
