% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runstan-data.R
\name{unlstan}
\alias{unlstan}
\title{unlstan}
\usage{
unlstan(stanfit, by1 = 1, start = 5000)
}
\arguments{
\item{stanfit}{results from call to stan}

\item{by1}{Skips for posterior draws}

\item{start}{Where to start by}
}
\description{
\code{unlstan} Unlist stan samples
}
