% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa-sim-fn.R
\name{geng}
\alias{geng}
\title{geng}
\usage{
geng(meansd, N, rmout = F, log = T)
}
\arguments{
\item{meansd}{Matrix with columns corresponding to source, mean, sd}

\item{N}{Number of days to simulate}

\item{rmout}{Whether to remove largest}

\item{log}{Whether lognormal}
}
\description{
\code{geng} Simulate G matrix
}
